FiveFor amount becomes generic and received qte by constructor
